Do not annotate plaintext renders
In most contexts we do not know how the plaintext
being rendered will be displayed, it can actually
ruin the presentation.

It is therefore safer to omit the annotations if
we know that the render is plaintext by design
(like in plain e-mails)

To make it work we also need a different way of
extracting the format from the render arguments.